Refactor Arabic researcher
