+++ packages/yoastseo/spec/morphology/arabic/stemmerCoverage/calculateCoverage.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-console */
-import stem from "../../../../src/languageProcessing/languages/ar/morphology/stem";
+import stem from "../../../../src/languageProcessing/languages/ar/helpers/internal/stem";
 import getMorphologyData from "../../../specHelpers/getMorphologyData";
 import goldStandard from "./goldStandardStems.json";
