+++ packages/yoastseo/spec/morphology/arabic/stemmerCoverage/generateStems.js
@@ -1,6 +1,6 @@
 /* eslint-disable no-console */
 import filterFunctionWordsFromArray from "../../../../src/languageProcessing/helpers/word/filterWordsFromArray";
-import stem from "../../../../src/languageProcessing/languages/ar/morphology/stem";
+import stem from "../../../../src/languageProcessing/languages/ar/helpers/internal/stem";
 import getMorphologyData from "../../../specHelpers/getMorphologyData";
 import sampleVocabulary from "../stemmerCoverage/sampleVocabulary.json";
