+++ packages/yoastseo/src/languageProcessing/languages/ar/Researcher.js
@@ -1,19 +1,15 @@
-import matchKeywordInSubheadings from "../../researches/base/matchKeywordInSubheadings.js";
-import getProminentWordsForInsights from "../../researches/base/getProminentWordsForInsights";
-import getProminentWordsForInternalLinking from "../../researches/base/getProminentWordsForInternalLinking";
-import getWordForms from "../../researches/_todo/getWordForms";
-import findKeywordInPageTitle from "../../researches/base/findKeywordInPageTitle";
-import { keyphraseDistributionResearcher as keyphraseDistribution } from "../../researches/base/keyphraseDistribution";
-
 import AbstractResearcher from "../../AbstractResearcher";
 
-import getLinkStatistics from "./researches/getLinkStatistics";
-import passiveVoice from "./researches/getPassiveVoice.js";
-import findTransitionWords from "./researches/findTransitionWords";
-import functionWordsInKeyphrase from "./researches/functionWordsInKeyphrase";
-import getSentenceBeginnings from "./researches/getSentenceBeginnings.js";
-import keyphraseLength from "./researches/keyphraseLength";
-import getKeywordDensity from "./researches/getKeywordDensity";
+// All config
+import firstWordExceptions from "./config/firstWordExceptions";
+import functionWords from "./config/functionWords";
+import transitionWords from "./config/transitionWords";
+import twoPartTransitionWords from "./config/twoPartTransitionWords";
+
+// All helpers
+import createBasicWordForms from "./helpers/createBasicWordForms";
+import getStemmer from "./helpers/getStemmer";
+import isPassiveSentence from "./helpers/isPassiveSentence";
 
 /**
  * The researches contains all the researches
@@ -27,20 +23,19 @@ export default class Researcher extends AbstractResearcher {
 	constructor( paper ) {
 		super( paper );
 
-		Object.assign( this.defaultResearches, {
-			findTransitionWords: findTransitionWords,
-			passiveVoice: passiveVoice,
-			getSentenceBeginnings: getSentenceBeginnings,
-			functionWordsInKeyphrase: functionWordsInKeyphrase,
-			matchKeywordInSubheadings: matchKeywordInSubheadings,
-			getLinkStatistics: getLinkStatistics,
-			keyphraseDistribution: keyphraseDistribution,
-			findKeywordInPageTitle: findKeywordInPageTitle,
-			morphology: getWordForms,
-			prominentWordsForInsights: getProminentWordsForInsights,
-			prominentWordsForInternalLinking: getProminentWordsForInternalLinking,
-			getKeywordDensity: getKeywordDensity,
-			keyphraseLength: keyphraseLength,
+		Object.assign( this.config, {
+			language: "ar",
+			isPeriphrastic: false,
+			firstWordExceptions,
+			functionWords,
+			transitionWords,
+			twoPartTransitionWords,
+		} );
+
+		Object.assign( this.helpers, {
+			createBasicWordForms,
+			getStemmer,
+			isPassiveSentence,
 		} );
 	}
 }
