+++ packages/yoastseo/src/languageProcessing/languages/ar/config/internal/passiveVerbsWithLongVowel.js
