+++ packages/yoastseo/src/languageProcessing/languages/ar/helpers/createBasicWordForms.js
@@ -5,7 +5,7 @@
  *
  * @returns {Array}        Prefixed and de-prefixed variations of a word.
  */
-export function createBasicWordForms( word ) {
+export default function createBasicWordForms( word ) {
 	const prefixes = [ "ل", "ب", "ال", "ك", "و", "ف", "س", "أ" ];
 	const forms = [];
