+++ packages/yoastseo/src/languageProcessing/languages/ar/helpers/internal/stem.js
